LayoutTests/fast/scrolling/scrollable-area-overflow-auto-visibility-visible.html

<!DOCTYPE html>
<head>
    <style type="text/css"> @import "resources/scrollable-style.css"; </style>
    <script src="../js/resources/js-test-pre.js"></script>
    <script src="resources/scrollable-area.js"></script>
</head>
<body onload="runTest(1);">
    <div id="scrollable" class="inner" style="background: red; overflow: auto; visibility: visible">content
        <div class="offscreen">content</div>
    </div>
    <div id='console'></div>
</body>
</html>
